+++ launch_chffrplus.sh
@@ -211,7 +211,7 @@ function launch {
 
   # handle pythonpath
   ln -sfn $(pwd) /data/pythonpath
-  export PYTHONPATH="$PWD"
+  export PYTHONPATH="$PWD:$PWD/pyextra"
 
   # hardware specific init
   if [ -f /EON ]; then