+++ selfdrive/common/util.cc
@@ -56,8 +56,8 @@ namespace util {
 std::string read_file(const std::string& fn) {
   std::ifstream ifs(fn, std::ios::binary | std::ios::ate);
   if (ifs) {
-    std::ifstream::pos_type pos = ifs.tellg();
-    if (pos != std::ios::beg) {
+    int pos = ifs.tellg();
+    if (pos > 0) {
       std::string result;
       result.resize(pos);
       ifs.seekg(0, std::ios::beg);
@@ -153,6 +153,10 @@ std::string dir_name(std::string const &path) {
   size_t pos = path.find_last_of("/");
   if (pos == std::string::npos) return "";
   return path.substr(0, pos);
+}
+
+bool is_valid_dongle_id(std::string const& dongle_id) {
+  return !dongle_id.empty() && dongle_id != "UnregisteredDevice";
 }
 
 struct tm get_time() {