+++ selfdrive/ui/qt/onroad.h
@@ -2,13 +2,16 @@
 
 #include <map>
 
+#include <QOpenGLFunctions>
+#include <QOpenGLWidget>
 #include <QSoundEffect>
-#include <QtWidgets>
+#include <QStackedLayout>
+#include <QWidget>
 
 #include "cereal/gen/cpp/log.capnp.h"
 #include "selfdrive/hardware/hw.h"
-#include "selfdrive/ui/ui.h"
 #include "selfdrive/ui/qt/qt_window.h"
+#include "selfdrive/ui/ui.h"
 
 typedef cereal::CarControl::HUDControl::AudibleAlert AudibleAlert;
 
@@ -29,21 +32,9 @@ private:
   void updateAlert(const QString &t1, const QString &t2, float blink_rate,
                    const std::string &type, cereal::ControlsState::AlertSize size, AudibleAlert sound);
 
-  std::map<AudibleAlert, std::pair<QString, bool>> sound_map {
-    // AudibleAlert, (file path, inf loop)
-    {AudibleAlert::CHIME_DISENGAGE, {"../assets/sounds/disengaged.wav", false}},
-    {AudibleAlert::CHIME_ENGAGE, {"../assets/sounds/engaged.wav", false}},
-    {AudibleAlert::CHIME_WARNING1, {"../assets/sounds/warning_1.wav", false}},
-    {AudibleAlert::CHIME_WARNING2, {"../assets/sounds/warning_2.wav", false}},
-    {AudibleAlert::CHIME_WARNING2_REPEAT, {"../assets/sounds/warning_2.wav", true}},
-    {AudibleAlert::CHIME_WARNING_REPEAT, {"../assets/sounds/warning_repeat.wav", true}},
-    {AudibleAlert::CHIME_ERROR, {"../assets/sounds/error.wav", false}},
-    {AudibleAlert::CHIME_PROMPT, {"../assets/sounds/error.wav", false}}
-  };
-
   QColor bg;
   float volume = Hardware::MIN_VOLUME;
-  std::map<AudibleAlert, QSoundEffect> sounds;
+  std::map<AudibleAlert, std::pair<QSoundEffect, int>> sounds;
   float blinking_rate = 0;
   QString text1, text2;
   std::string alert_type;