+++ selfdrive/ui/qt/request_repeater.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "selfdrive/common/util.h"
 #include "selfdrive/ui/qt/api.h"
 #include "selfdrive/ui/ui.h"
 
@@ -8,5 +9,7 @@ public:
   RequestRepeater(QObject *parent, const QString &requestURL, const QString &cacheKey = "", int period = 0);
 
 private:
+  Params params;
   QTimer *timer;
+  QString prevResp;
 };