+++ selfdrive/ui/qt/sidebar.cc
@@ -1,5 +1,7 @@
 #include "selfdrive/ui/qt/sidebar.h"
 
+#include <QMouseEvent>
+
 #include "selfdrive/ui/qt/qt_window.h"
 #include "selfdrive/common/util.h"
 #include "selfdrive/hardware/hw.h"
@@ -55,7 +57,8 @@ void Sidebar::updateState(const UIState &s) {
 
   auto deviceState = sm["deviceState"].getDeviceState();
   setProperty("netType", network_type[deviceState.getNetworkType()]);
-  setProperty("netStrength", (int)deviceState.getNetworkStrength());
+  int strength = (int)deviceState.getNetworkStrength();
+  setProperty("netStrength", strength > 0 ? strength + 1 : 0);
 
   auto last_ping = deviceState.getLastAthenaPingTime();
   if (last_ping == 0) {