+++ selfdrive/ui/qt/widgets/controls.h
@@ -4,7 +4,6 @@
 #include <QHBoxLayout>
 #include <QLabel>
 #include <QPushButton>
-#include <QVBoxLayout>
 
 #include "selfdrive/common/params.h"
 #include "selfdrive/ui/qt/widgets/toggle.h"
@@ -57,32 +56,15 @@ class ButtonControl : public AbstractControl {
   Q_OBJECT
 
 public:
-  template <typename Functor>
-  ButtonControl(const QString &title, const QString &text, const QString &desc, Functor functor, const QString &icon = "", QWidget *parent = nullptr) : AbstractControl(title, desc, icon, parent) {
-    btn.setText(text);
-    btn.setStyleSheet(R"(
-      QPushButton {
-        padding: 0;
-        border-radius: 50px;
-        font-size: 35px;
-        font-weight: 500;
-        color: #E4E4E4;
-        background-color: #393939;
-      }
-      QPushButton:disabled {
-        color: #33E4E4E4;
-      }
-    )");
-    btn.setFixedSize(250, 100);
-    QObject::connect(&btn, &QPushButton::released, functor);
-    hlayout->addWidget(&btn);
-  }
-  void setText(const QString &text) { btn.setText(text); }
+  ButtonControl(const QString &title, const QString &text, const QString &desc = "", QWidget *parent = nullptr);
+  inline void setText(const QString &text) { btn.setText(text); }
+  inline QString text() const { return btn.text(); }
+
+signals:
+  void released();
 
 public slots:
-  void setEnabled(bool enabled) {
-    btn.setEnabled(enabled);
-  };
+  void setEnabled(bool enabled) { btn.setEnabled(enabled); };
 
 private:
   QPushButton btn;