+++ RELEASES.md
@@ -2,6 +2,8 @@ Version 0.8.6 (2021-XX-XX)
 ========================
  * Hyundai Ioniq PHEV 2020 support thanks to YawWashout!
  * Škoda Octavia RS 2016 support thanks to jyoung8607!
+ * Volkswagen Golf SportWagen 2015 support thanks to jona96!
+ * Volkswagen Touran 2017 support thanks to jyoung8607!
 
 Version 0.8.5 (2021-06-11)
 ========================